+++ CHANGELOG.md
@@ -40,6 +40,7 @@ omnibus-gitlab repository.
 
 - Make default IMAP incoming mailbox "inbox" in case user omits this setting d3c187
 - Make NGINX server_names_hash_bucket_size configurable and default it to 64 bytes 7cb488
+- Use gitlab:db:configure to seed and migrate the database
 - Add log prefix for pages and registry services 48e29b
 - Add configuration option for the Container Registry storage driver
 - Change the autovacuum configuration defaults f5ac85